bot only deletes its most recent vc

+++ bot.py
@@ -28,13 +28,18 @@ async def on_ready():
     for key in commands:
         print(commands[key].name)
 
+
+#  ---------------Events-------------------
+
 @bot.event
 async def on_voice_state_update(member, before, after):
-  if before.channel is not None:
+  from behaviors.createVC import NewId
+  if before.channel is not None: 
    if before.channel.id != 610818618325729285: # DO NOT REMOVE!!!!!
-    if len(before.channel.members) == 0:
-      print("channel is now empty")
-      await before.channel.delete()
+     if before.channel.id == NewId:
+      if len(before.channel.members) == 0:
+        print("channel is now empty")
+        await before.channel.delete()
